requirements.txt
asgiref==3.5.0
astroid==2.5.2
autopep8==1.5.7
certifi==2021.5.30
chardet==4.0.0
colorama==0.4.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2
django-admin-rangefilter==0.8.1
django-auditlog==2.1.1
django-ckeditor==6.1.0
django-cors-headers==3.7.0
django-db-prefix==1.0.5
django-extensions==3.1.5
django-filter==2.4.0
django-import-export==2.7.1
django-js-asset==1.2.2
django-middleware-global-request==0.3.0
django-mysql==4.5.0
django-phonenumber-field==5.0.0
djangorestframework==3.11.2
djangorestframework-simplejwt==4.7.0
dynamic-rest==2.1.1
et-xmlfile==1.1.0
gunicorn==20.1.0
hashids==1.3.1
idna==2.10
importlib-metadata==4.6.1
inflection==0.5.1
isort==5.8.0
lazy-object-proxy==1.6.0
Markdown==3.3.4
MarkupPy==1.14
mccabe==0.6.1
mssql-django==1.1.1
odfpy==1.4.1
openpyxl==3.0.9
phonenumbers==8.12.28
phonenumberslite==8.12.20
Pillow==8.4.0
psycopg2-binary==2.8.6
pycodestyle==2.7.0
pydotplus==2.0.2
PyJWT==2.1.0
pylint==2.7.4
PyMySQL==1.0.2
pyodbc==4.0.32
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.1
PyYAML==6.0
reportlab==3.6.9
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
tablib==3.2.0
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.4
whitenoise==5.2.0
wrapt==1.12.1
xlrd==2.0.1
XlsxWriter==3.0.3
xlwt==1.3.0
zipp==3.5.0